Fix decimal smoke readings and realtime chart

+++ backend/src/main/java/com/smoke/dto/CurrentReadingResponse.java
@@ -1,12 +1,13 @@
 package com.smoke.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CurrentReadingResponse(
         Long id,
         String deviceId,
         String deviceName,
-        Integer concentration,
+        BigDecimal concentration,
         LocalDateTime timestamp,
         Integer threshold,
         boolean online) {
